requirements.txt
PyOpenGL==3.1.7
numpy==1.24.3
